refactor: move googleauthenticator transformer to layout dir
